+++ Projects/Game/Player.cpp
@@ -390,8 +390,8 @@ void Player::EntryDraw()
 		DrawString(Game::kScreenWidth / 2, Game::kScreenHeight / 2, L"Ready、", 0xff0000);
 	}
 
-	int drawX = m_pos.x - m_pCamera->scroll.x;
-	int drawY = m_pos.y - m_pCamera->scroll.y;
+	int drawX = static_cast<int>(m_pos.x - m_pCamera->scroll.x);
+	int drawY = static_cast<int>(m_pos.y - m_pCamera->scroll.y);
 	DrawBox(drawX - 16, drawY - 16, drawX + 16, drawY + 16, 0x22ff00, true);
 	DrawRectRotaGraph(drawX, drawY, 0, 0, kGraphWidth, kGraphHeight, 1.0f, 0.0f, m_graphHandle, true, false);
 
@@ -403,8 +403,8 @@ void Player::EntryDraw()
 
 void Player::JumpDraw()
 {
-	int drawX = m_pos.x - m_pCamera->scroll.x;
-	int drawY = m_pos.y - m_pCamera->scroll.y;
+	int drawX = static_cast<int>(m_pos.x - m_pCamera->scroll.x);
+	int drawY = static_cast<int>(m_pos.y - m_pCamera->scroll.y);
 	DrawBox(drawX - 16, drawY - 16, drawX + 16, drawY + 16, 0xaaffff, true);
 	DrawRectRotaGraph(drawX, drawY, kGraphWidth * 3, 0, kGraphWidth, kGraphHeight, 1.0f, 0.0f, m_graphHandle, true, false);
 
@@ -417,8 +417,8 @@ void Player::JumpDraw()
 
 void Player::GroundDraw()
 {
-	int drawX = m_pos.x - m_pCamera->scroll.x;
-	int drawY = m_pos.y - m_pCamera->scroll.y;
+	int drawX = static_cast<int>(m_pos.x - m_pCamera->scroll.x);
+	int drawY = static_cast<int>(m_pos.y - m_pCamera->scroll.y);
 	DrawBox(drawX - 16, drawY - 16, drawX + 16, drawY + 16, 0x22ff00, true);
 	DrawRectRotaGraph(drawX, drawY, 0, 0, kGraphWidth, kGraphHeight, 1.0f, 0.0f, m_graphHandle, true, false);
 	
@@ -430,8 +430,8 @@ void Player::GroundDraw()
 
 void Player::MissDraw()
 {
-	int drawX = m_pos.x - m_pCamera->scroll.x;
-	int drawY = m_pos.y - m_pCamera->scroll.y;
+	int drawX = static_cast<int>(m_pos.x - m_pCamera->scroll.x);
+	int drawY = static_cast<int>(m_pos.y - m_pCamera->scroll.y);
 	DrawBox(drawX - 16, drawY - 16, drawX + 16, drawY + 16, 0xff00aa, true);
 	DrawRectRotaGraph(drawX, drawY, kGraphWidth, 0, kGraphWidth, kGraphHeight, 1.0f, 0.0f, m_graphHandle, true, false);
